+++ examples/ERP/plot_embedding_MEG.py
@@ -1,17 +1,31 @@
 """
-=====================================================================
-Embedding ERP MEG data in 2D Euclidean space with Laplacian Eigenmaps
-=====================================================================
+============================================
+Embedding ERP MEG data in 2D Euclidean space
+============================================
 
-Spectral embedding via Laplacian Eigenmaps of a set of ERP data.
+Embeddings via Laplacian Eigenmaps and Riemannian locally linear
+embedding of a set of ERP data. Embedding via Laplacian Eigenmaps is
+referred to as Spectral Embedding (SE).
 
+"Locally Linear Embedding (LLE) assumes that the  local neighborhood of a
+point on the manifold can be well approximated by  the affine subspace
+spanned by the k-nearest neighbors of the point and finds a low-dimensional
+embedding of the data based on these affine approximations. Laplacian
+Eigenmaps (LE) are based on computing the low dimensional representation
+that best preserves locality instead of local linearity in LLE." [1]_
+
+References
+----------
+.. [1] A. Goh and R. Vidal, "Clustering and dimensionality reduction
+    on Riemannian manifolds", 2008 IEEE Conference on Computer Vision
+    and Pattern Recognition, June 2008.
 """
-# Authors: Pedro Rodrigues <pedro.rodrigues01@gmail.com>
-#
+# Authors:  Pedro Rodrigues <pedro.rodrigues01@gmail.com>,
+#           Gabriel Wagner vom Berg <gabriel@bccn-berlin.de>
 # License: BSD (3-clause)
 
 from pyriemann.estimation import XdawnCovariances
-from pyriemann.embedding import Embedding
+from pyriemann.utils.viz import plot_embedding
 
 import mne
 from mne import io
@@ -19,7 +33,7 @@ from mne.datasets import sample
 
 from sklearn.model_selection import train_test_split
 
-from matplotlib import pyplot as plt
+import matplotlib.pyplot as plt
 
 
 print(__doc__)
@@ -50,7 +64,7 @@ X = epochs.get_data()
 y = epochs.events[:, -1]
 
 ###############################################################################
-# Embedding the Xdawn covariance matrices with Laplacian Eigenmaps
+# Embedding of Xdawn covariance matrices
 
 nfilter = 4
 xdwn = XdawnCovariances(estimator='scm', nfilter=nfilter)
@@ -58,23 +72,18 @@ split = train_test_split(X, y, train_size=0.25, random_state=42)
 Xtrain, Xtest, ytrain, ytest = split
 covs = xdwn.fit(Xtrain, ytrain).transform(Xtest)
 
-lapl = Embedding(metric='riemann', n_components=2)
-embd = lapl.fit_transform(covs)
+###############################################################################
+# Spectral Embedding (SE)
+# -----------------------
+
+plot_embedding(covs, ytest, metric='riemann', embd_type='Spectral',
+               normalize=True)
+plt.show()
 
 ###############################################################################
-# Plot the three first components of the embedded points
+# Riemannian Locally Linear Embedding (RLLE)
+# ------------------------------------------
 
-fig, ax = plt.subplots(figsize=(7, 8), facecolor='white')
-
-for cond, label in event_id.items():
-    idx = (ytest == label)
-    ax.scatter(embd[idx, 0], embd[idx, 1], s=36, label=cond)
-
-ax.set_xlabel(r'$\varphi_1$', fontsize=16)
-ax.set_ylabel(r'$\varphi_2$', fontsize=16)
-ax.set_title('Spectral embedding of ERP recordings', fontsize=16)
-ax.set_xticks([-1.0, -0.5, 0.0, +0.5, 1.0])
-ax.set_yticks([-1.0, -0.5, 0.0, +0.5, 1.0])
-ax.grid(False)
-ax.legend()
+plot_embedding(covs, ytest, metric='riemann', embd_type='LocallyLinear',
+               normalize=False)
 plt.show()